// FrontEnd/src/app/model/systemFunctionOutputModel.ts
/**
 * 走到飛 API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * 大功能OutputModel
 */
export interface SystemFunctionOutputModel { 
    /**
     * 大功能Id
     */
    functionId?: number;
    /**
     * 大功能名稱
     */
    functionName?: string;
    /**
     * 大功能中文名稱
     */
    functionChineseName?: string;
    /**
     * 是否啟用
     */
    enableFlag?: boolean;
    /**
     * 備註
     */
    remark?: string;
    /**
     * 排序
     */
    sort?: number;
}